Move files around to isolate game definitions
