+++ packages/api/src/availableServices.ts
@@ -3,11 +3,11 @@ import { HealthServiceApi, HealthServiceDefinition } from "./services/health";
 import {
   TileGameServiceApi,
   TileGameServiceDefinition
-} from "./services/tileGame";
+} from "./services/games/tiles-tbd/tileGame/tileGame";
 import {
   TileMapServiceApi,
   TileMapServiceDefinition
-} from "./services/tileMap";
+} from "./services/games/tiles-tbd/tileMap/tileMap";
 
 export type AvailableServices = {
   health: HealthServiceApi;
