+++ packages/api/src/genericTypes/index.ts
@@ -1,0 +1,4 @@
+export * from "./error";
+export * from "./removeExtendsString";
+export * from "./service";
+export * from "./socket";
