+++ packages/api/src/index.ts
@@ -1,15 +1,6 @@
-export * from "./genericTypes/error";
-export * from "./genericTypes/removeExtendsString";
-export * from "./genericTypes/service";
-export * from "./genericTypes/socket";
-
-export * from "./services/tileGame/types";
-export * from "./services/tileMap/types";
-export * from "./services/tileSocket/types";
+export * from "./genericTypes";
 
 export * from "./services/health";
-export * from "./services/tileGame";
-export * from "./services/tileMap";
-export * from "./services/tileSocket";
+export * from "./services/games";
 
 export * from "./availableServices";
