+++ packages/api/src/services/games/index.ts
@@ -1,0 +1,1 @@
+export * from "./tiles-tbd";
