+++ packages/api/src/services/games/tiles-tbd/index.ts
@@ -1,0 +1,3 @@
+export * from "./tileGame";
+export * from "./tileMap";
+export * from "./tileSocket";
