+++ packages/api/src/services/games/tiles-tbd/tileGame/index.ts
@@ -1,0 +1,2 @@
+export * from "./tileGame";
+export * from "./types";
