+++ packages/api/src/services/games/tiles-tbd/tileMap/index.ts
@@ -1,0 +1,2 @@
+export * from "./tileMap";
+export * from "./types";
