+++ packages/api/src/services/games/tiles-tbd/tileSocket/index.ts
@@ -1,0 +1,2 @@
+export * from "./tileSocket";
+export * from "./types";
