+++ packages/backend/src/games/games.module.ts
@@ -1,0 +1,7 @@
+import { Module } from "@nestjs/common";
+import { TilesTbdModule } from "./tiles-tbd/tilesTbd.module";
+
+@Module({
+  imports: [TilesTbdModule]
+})
+export class GamesModule {}
