+++ packages/backend/src/games/tiles-tbd/map/tileMap.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common";
-import { PrismaModule } from "src/database/prisma.module";
+import { PrismaModule } from "src/games/tiles-tbd/database/prisma.module";
 import { TileMapController } from "./tileMap.controller";
 import { TileMapService } from "./tileMap.service";
