+++ packages/backend/src/games/tiles-tbd/map/tileMap.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from "@nestjs/common";
 import { CompleteTileMap, TileMap, TileMapId } from "@tiles-tbd/api";
-import { PrismaService } from "src/database/prisma.service";
-import { BaseTileMapGenerator } from "src/generators/baseGenerator";
-import { MagicMazeLikeMapGenerator as MagicMazeSimpleTileMapGenerator } from "src/generators/gridMapGenerator";
+import { PrismaService } from "src/games/tiles-tbd/database/prisma.service";
+import { BaseTileMapGenerator } from "src/games/tiles-tbd/generators/baseGenerator";
+import { MagicMazeLikeMapGenerator as MagicMazeSimpleTileMapGenerator } from "src/games/tiles-tbd/generators/gridMapGenerator";
 
 const AVAILABLE_TILE_MAP_GENERATORS: Record<
   string,
