+++ packages/backend/src/games/tiles-tbd/socket/tileSocket.gateway.ts
@@ -1,9 +1,6 @@
-import { Logger } from "@nestjs/common";
 import {
   ConnectedSocket,
   MessageBody,
-  OnGatewayConnection,
-  OnGatewayDisconnect,
   WebSocketGateway,
   WebSocketServer
 } from "@nestjs/websockets";
@@ -20,37 +17,20 @@ import {
   getSocketEmitter,
   SocketGatewayHandleMessage
 } from "src/genericTypes/socket";
+import { SocketGateway } from "src/genericTypes/socket.gateway";
 
 @WebSocketGateway({
   cors: {
     origin: "*"
   }
 })
-export class SocketGateway
-  implements
-    OnGatewayConnection,
-    OnGatewayDisconnect,
-    SocketGatewayHandleMessage<TileFromClientToServer>
+export class TileSocketGateway
+  extends SocketGateway
+  implements SocketGatewayHandleMessage<TileFromClientToServer>
 {
-  private logger = new Logger("Socket");
   public clients: Map<string, Socket> = new Map();
 
   @WebSocketServer() server: Server;
-
-  afterInit() {
-    this.logger.log("WebSocket Gateway initialized");
-  }
-
-  handleConnection(client: Socket) {
-    this.logger.log(`Client connected: ${client.id}`);
-
-    this.clients.set(client.id, client);
-  }
-
-  handleDisconnect(client: Socket) {
-    this.logger.log(`Client disconnected: ${client.id}`);
-    this.clients.delete(client.id);
-  }
 
   @getSocketDecorator(TileServerSocketDefinition.receiveMessage.identify)
   identify(
