+++ packages/backend/src/games/tiles-tbd/tilesTbd.module.ts
@@ -1,0 +1,8 @@
+import { Module } from "@nestjs/common";
+import { TileMapModule } from "./map/tileMap.module";
+import { TileGameModule } from "./game/tileGame.module";
+
+@Module({
+  imports: [TileMapModule, TileGameModule]
+})
+export class TilesTbdModule {}
