+++ packages/backend/src/genericTypes/socket.gateway.ts
@@ -1,0 +1,35 @@
+import { Logger } from "@nestjs/common";
+import {
+  OnGatewayConnection,
+  OnGatewayDisconnect,
+  WebSocketGateway,
+  WebSocketServer
+} from "@nestjs/websockets";
+import { Server, Socket } from "socket.io";
+
+@WebSocketGateway({
+  cors: {
+    origin: "*"
+  }
+})
+export class SocketGateway implements OnGatewayConnection, OnGatewayDisconnect {
+  public logger = new Logger("Socket");
+  public clients: Map<string, Socket> = new Map();
+
+  @WebSocketServer() server: Server;
+
+  afterInit() {
+    this.logger.log("WebSocket Gateway initialized");
+  }
+
+  handleConnection(client: Socket) {
+    this.logger.log(`Client connected: ${client.id}`);
+
+    this.clients.set(client.id, client);
+  }
+
+  handleDisconnect(client: Socket) {
+    this.logger.log(`Client disconnected: ${client.id}`);
+    this.clients.delete(client.id);
+  }
+}
